backend/requirements.txt
fastapi 
uvicorn[standard]
python-multipart
tensorflow[and-cuda]
typing-extensions==4.8.0
tqdm
Pillow
blurhash